requirements.txt
pycrypto>=2.6.1
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
gunicorn==19.7.1
PyMySQL==0.8.0
